formatting and bigger boxes

+++ src/main/java/planitpoker/voting/CardsPanel.java
@@ -5,7 +5,7 @@ import java.util.ArrayList;
 import javax.swing.JPanel;
 
 public class CardsPanel extends JPanel {
-    public CardsPanel(VotingController votingController, ArrayList<Integer> cards) {
+    public CardsPanel(VotingController votingController, ArrayList<Double> cards) {
         super();
 
         GridLayout layout = new GridLayout(2,3);
@@ -13,7 +13,7 @@ public class CardsPanel extends JPanel {
         layout.setVgap(10);
         setLayout(layout);
 
-        for (int card : cards) {
+        for (double card : cards) {
             add(new CardPanel(votingController, card));
         }
     }
